Rename `Whitespace` to `TrailingWhitespace`
