+++ lib/git_hooks/pre_commit.rb
@@ -2,7 +2,7 @@ require_relative 'pre_commit/prevent_debugger'
 require_relative 'pre_commit/prevent_master'
 require_relative 'pre_commit/rspec'
 require_relative 'pre_commit/rubocop'
-require_relative 'pre_commit/whitespace'
+require_relative 'pre_commit/trailing_whitespace'
 
 module GitHooks
   module PreCommit
