+++ lib/git_hooks/trailing_whitespace_validator.rb
@@ -1,5 +1,5 @@
 module GitHooks
-  class WhitespaceValidator
+  class TrailingWhitespaceValidator
     def errors?(files)
       files
         .map { |file| File.open(file) }
